ch(tests):write more test of endpionts [Finishes #164800778]
